Minor Tsavorite reorg: Break the various contexts into their own files, and move RecoveryInfo into CheckpointManagement.
