+++ libs/storage/Tsavorite/cs/benchmark/FixedLenYcsbBenchmark.cs
@@ -1,9 +1,6 @@
 ﻿// Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 
-// Define below to enable continuous performance report for dashboard
-// #define DASHBOARD
-
 using System;
 using System.Diagnostics;
 using System.Runtime.InteropServices;
@@ -53,20 +50,6 @@ namespace Tsavorite.benchmark
             upsertPercent = testLoader.UpsertPercent;
             rmwPercent = testLoader.RmwPercent;
             functions = new Functions();
-
-#if DASHBOARD
-            statsWritten = new AutoResetEvent[threadCount];
-            for (int i = 0; i < threadCount; i++)
-            {
-                statsWritten[i] = new AutoResetEvent(false);
-            }
-            threadThroughput = new double[threadCount];
-            threadAverageLatency = new double[threadCount];
-            threadMaximumLatency = new double[threadCount];
-            threadProgress = new long[threadCount];
-            writeStats = new bool[threadCount];
-            freq = Stopwatch.Frequency;
-#endif
 
             input_ = new Input[8];
             for (int i = 0; i < 8; i++)
@@ -132,13 +115,6 @@ namespace Tsavorite.benchmark
             long writes_done = 0;
             long deletes_done = 0;
 
-#if DASHBOARD
-            var tstart = Stopwatch.GetTimestamp();
-            var tstop1 = tstart;
-            var lastWrittenValue = 0;
-            int count = 0;
-#endif
-
             using var session = store.NewSession<Input, Output, Empty, Functions>(functions);
             var uContext = session.UnsafeContext;
             uContext.BeginUnsafe();
@@ -185,22 +161,6 @@ namespace Tsavorite.benchmark
                         uContext.Delete(ref txn_keys_[idx], Empty.Default);
                         ++deletes_done;
                     }
-
-#if DASHBOARD
-                    count += (int)kChunkSize;
-
-                    //Check if stats collector is requesting for statistics
-                    if (writeStats[thread_idx])
-                    {
-                        var tstart1 = tstop1;
-                        tstop1 = Stopwatch.GetTimestamp();
-                        threadProgress[thread_idx] = count;
-                        threadThroughput[thread_idx] = (count - lastWrittenValue) / ((tstop1 - tstart1) / freq);
-                        lastWrittenValue = count;
-                        writeStats[thread_idx] = false;
-                        statsWritten[thread_idx].Set();
-                    }
-#endif
                 }
 
                 uContext.CompletePending(true);
@@ -211,10 +171,6 @@ namespace Tsavorite.benchmark
             }
 
             sw.Stop();
-
-#if DASHBOARD
-            statsWritten[thread_idx].Set();
-#endif
 
             Console.WriteLine($"Thread {thread_idx} done; {reads_done} reads, {writes_done} writes, {deletes_done} deletes in {sw.ElapsedMilliseconds} ms.");
             Interlocked.Add(ref total_ops_done, reads_done + writes_done + deletes_done);
@@ -295,11 +251,6 @@ namespace Tsavorite.benchmark
 
         internal unsafe (double insPerSec, double opsPerSec, long tailAddress) Run(TestLoader testLoader)
         {
-#if DASHBOARD
-            var dash = new Thread(() => DoContinuousMeasurements());
-            dash.Start();
-#endif
-
             Thread[] workers = new Thread[testLoader.Options.ThreadCount];
 
             Console.WriteLine("Executing setup.");
@@ -320,16 +271,13 @@ namespace Tsavorite.benchmark
                 }
 
                 foreach (Thread worker in workers)
-                {
                     worker.Start();
-                }
 
                 waiter.Set();
                 var sw = Stopwatch.StartNew();
                 foreach (Thread worker in workers)
-                {
                     worker.Join();
-                }
+
                 sw.Stop();
                 elapsedMs = sw.ElapsedMilliseconds;
                 waiter.Reset();
@@ -366,9 +314,7 @@ namespace Tsavorite.benchmark
             }
             // Start threads.
             foreach (Thread worker in workers)
-            {
                 worker.Start();
-            }
 
             waiter.Set();
             var swatch = Stopwatch.StartNew();
@@ -400,17 +346,10 @@ namespace Tsavorite.benchmark
             swatch.Stop();
 
             done = true;
-
             foreach (Thread worker in workers)
-            {
                 worker.Join();
-            }
 
             waiter.Reset();
-
-#if DASHBOARD
-            dash.Join();
-#endif
 
             double seconds = swatch.ElapsedMilliseconds / 1000.0;
             Console.WriteLine(TestStats.GetAddressesLine(AddressLineNum.After, store.Log.BeginAddress, store.Log.HeadAddress, store.Log.ReadOnlyAddress, store.Log.TailAddress));
@@ -436,13 +375,6 @@ namespace Tsavorite.benchmark
             var uContext = session.UnsafeContext;
             uContext.BeginUnsafe();
 
-#if DASHBOARD
-            var tstart = Stopwatch.GetTimestamp();
-            var tstop1 = tstart;
-            var lastWrittenValue = 0;
-            int count = 0;
-#endif
-
             Value value = default;
 
             try
@@ -465,20 +397,6 @@ namespace Tsavorite.benchmark
 
                         uContext.Upsert(ref init_keys_[idx], ref value, Empty.Default);
                     }
-#if DASHBOARD
-                count += (int)kChunkSize;
-
-                //Check if stats collector is requesting for statistics
-                if (writeStats[thread_idx])
-                {
-                    var tstart1 = tstop1;
-                    tstop1 = Stopwatch.GetTimestamp();
-                    threadThroughput[thread_idx] = (count - lastWrittenValue) / ((tstop1 - tstart1) / freq);
-                    lastWrittenValue = count;
-                    writeStats[thread_idx] = false;
-                    statsWritten[thread_idx].Set();
-                }
-#endif
                 }
                 uContext.CompletePending(true);
             }
@@ -527,70 +445,6 @@ namespace Tsavorite.benchmark
 
             bContext.CompletePending(true);
         }
-
-#if DASHBOARD
-        int measurementInterval = 2000;
-        bool measureLatency;
-        bool[] writeStats;
-        private EventWaitHandle[] statsWritten;
-        double[] threadThroughput;
-        double[] threadAverageLatency;
-        double[] threadMaximumLatency;
-        long[] threadProgress;
-        double freq;
-
-        void DoContinuousMeasurements()
-        {
-            double totalThroughput, totalLatency, maximumLatency;
-            double totalProgress;
-            int ver = 0;
-
-            using (var client = new WebClient())
-            {
-                while (!done)
-                {
-                    ver++;
-
-                    Thread.Sleep(measurementInterval);
-
-                    totalProgress = 0;
-                    totalThroughput = 0;
-                    totalLatency = 0;
-                    maximumLatency = 0;
-
-                    for (int i = 0; i < threadCount; i++)
-                    {
-                        writeStats[i] = true;
-                    }
-
-
-                    for (int i = 0; i < threadCount; i++)
-                    {
-                        statsWritten[i].WaitOne();
-                        totalThroughput += threadThroughput[i];
-                        totalProgress += threadProgress[i];
-                        if (measureLatency)
-                        {
-                            totalLatency += threadAverageLatency[i];
-                            if (threadMaximumLatency[i] > maximumLatency)
-                            {
-                                maximumLatency = threadMaximumLatency[i];
-                            }
-                        }
-                    }
-
-                    if (measureLatency)
-                    {
-                        Console.WriteLine("{0} \t {1:0.000} \t {2} \t {3} \t {4} \t {5}", ver, totalThroughput / (double)1000000, totalLatency / threadCount, maximumLatency, store.Log.TailAddress, totalProgress);
-                    }
-                    else
-                    {
-                        Console.WriteLine("{0} \t {1:0.000} \t {2} \t {3}", ver, totalThroughput / (double)1000000, store.Log.TailAddress, totalProgress);
-                    }
-                }
-            }
-        }
-#endif
 
         #region Load Data
