+++ libs/storage/Tsavorite/cs/benchmark/SpanByteYcsbBenchmark.cs
@@ -1,9 +1,6 @@
 ﻿// Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 
-// Define below to enable continuous performance report for dashboard
-// #define DASHBOARD
-
 using System;
 using System.Diagnostics;
 using System.Runtime.InteropServices;
@@ -56,20 +53,6 @@ namespace Tsavorite.benchmark
             upsertPercent = testLoader.UpsertPercent;
             rmwPercent = testLoader.RmwPercent;
             functions = new FunctionsSB();
-
-#if DASHBOARD
-            statsWritten = new AutoResetEvent[threadCount];
-            for (int i = 0; i < threadCount; i++)
-            {
-                statsWritten[i] = new AutoResetEvent(false);
-            }
-            threadThroughput = new double[threadCount];
-            threadAverageLatency = new double[threadCount];
-            threadMaximumLatency = new double[threadCount];
-            threadProgress = new long[threadCount];
-            writeStats = new bool[threadCount];
-            freq = Stopwatch.Frequency;
-#endif
 
             input_ = new Input[8];
             for (int i = 0; i < 8; i++)
@@ -145,13 +128,6 @@ namespace Tsavorite.benchmark
             long writes_done = 0;
             long deletes_done = 0;
 
-#if DASHBOARD
-            var tstart = Stopwatch.GetTimestamp();
-            var tstop1 = tstart;
-            var lastWrittenValue = 0;
-            int count = 0;
-#endif
-
             using var session = store.NewSession<SpanByte, SpanByteAndMemory, Empty, FunctionsSB>(functions);
             var uContext = session.UnsafeContext;
             uContext.BeginUnsafe();
@@ -198,22 +174,6 @@ namespace Tsavorite.benchmark
                         uContext.Delete(ref SpanByte.Reinterpret(ref txn_keys_[idx]), Empty.Default);
                         ++deletes_done;
                     }
-
-#if DASHBOARD
-                    count += (int)kChunkSize;
-
-                    //Check if stats collector is requesting for statistics
-                    if (writeStats[thread_idx])
-                    {
-                        var tstart1 = tstop1;
-                        tstop1 = Stopwatch.GetTimestamp();
-                        threadProgress[thread_idx] = count;
-                        threadThroughput[thread_idx] = (count - lastWrittenValue) / ((tstop1 - tstart1) / freq);
-                        lastWrittenValue = count;
-                        writeStats[thread_idx] = false;
-                        statsWritten[thread_idx].Set();
-                    }
-#endif
                 }
 
                 uContext.CompletePending(true);
@@ -224,10 +184,6 @@ namespace Tsavorite.benchmark
             }
 
             sw.Stop();
-
-#if DASHBOARD
-            statsWritten[thread_idx].Set();
-#endif
 
             Console.WriteLine($"Thread {thread_idx} done; {reads_done} reads, {writes_done} writes, {deletes_done} deletes in {sw.ElapsedMilliseconds} ms.");
             Interlocked.Add(ref total_ops_done, reads_done + writes_done + deletes_done);
@@ -259,13 +215,6 @@ namespace Tsavorite.benchmark
             long reads_done = 0;
             long writes_done = 0;
             long deletes_done = 0;
-
-#if DASHBOARD
-            var tstart = Stopwatch.GetTimestamp();
-            var tstop1 = tstart;
-            var lastWrittenValue = 0;
-            int count = 0;
-#endif
 
             using var session = store.NewSession<SpanByte, SpanByteAndMemory, Empty, FunctionsSB>(functions);
             var bContext = session.BasicContext;
@@ -311,31 +260,11 @@ namespace Tsavorite.benchmark
                     bContext.Delete(ref SpanByte.Reinterpret(ref txn_keys_[idx]), Empty.Default);
                     ++deletes_done;
                 }
-
-#if DASHBOARD
-                count += (int)kChunkSize;
-
-                //Check if stats collector is requesting for statistics
-                if (writeStats[thread_idx])
-                {
-                    var tstart1 = tstop1;
-                    tstop1 = Stopwatch.GetTimestamp();
-                    threadProgress[thread_idx] = count;
-                    threadThroughput[thread_idx] = (count - lastWrittenValue) / ((tstop1 - tstart1) / freq);
-                    lastWrittenValue = count;
-                    writeStats[thread_idx] = false;
-                    statsWritten[thread_idx].Set();
-                }
-#endif
             }
 
             bContext.CompletePending(true);
 
             sw.Stop();
-
-#if DASHBOARD
-            statsWritten[thread_idx].Set();
-#endif
 
             Console.WriteLine($"Thread {thread_idx} done; {reads_done} reads, {writes_done} writes, {deletes_done} deletes in {sw.ElapsedMilliseconds} ms.");
             Interlocked.Add(ref total_ops_done, reads_done + writes_done + deletes_done);
@@ -343,11 +272,6 @@ namespace Tsavorite.benchmark
 
         internal unsafe (double insPerSec, double opsPerSec, long tailAddress) Run(TestLoader testLoader)
         {
-#if DASHBOARD
-            var dash = new Thread(() => DoContinuousMeasurements());
-            dash.Start();
-#endif
-
             Thread[] workers = new Thread[testLoader.Options.ThreadCount];
 
             Console.WriteLine("Executing setup.");
@@ -368,16 +292,13 @@ namespace Tsavorite.benchmark
                 }
 
                 foreach (Thread worker in workers)
-                {
                     worker.Start();
-                }
 
                 waiter.Set();
                 var sw = Stopwatch.StartNew();
                 foreach (Thread worker in workers)
-                {
                     worker.Join();
-                }
+
                 sw.Stop();
                 waiter.Reset();
 
@@ -413,11 +334,10 @@ namespace Tsavorite.benchmark
                 else
                     workers[idx] = new Thread(() => RunYcsbUnsafeContext(x));
             }
+
             // Start threads.
             foreach (Thread worker in workers)
-            {
                 worker.Start();
-            }
 
             waiter.Set();
             var swatch = Stopwatch.StartNew();
@@ -449,16 +369,10 @@ namespace Tsavorite.benchmark
             swatch.Stop();
 
             done = true;
-
             foreach (Thread worker in workers)
-            {
                 worker.Join();
-            }
-            waiter.Reset();
 
-#if DASHBOARD
-            dash.Join();
-#endif
+            waiter.Reset();
 
             double seconds = swatch.ElapsedMilliseconds / 1000.0;
             Console.WriteLine(TestStats.GetAddressesLine(AddressLineNum.After, store.Log.BeginAddress, store.Log.HeadAddress, store.Log.ReadOnlyAddress, store.Log.TailAddress));
@@ -484,13 +398,6 @@ namespace Tsavorite.benchmark
             var uContext = session.UnsafeContext;
             uContext.BeginUnsafe();
 
-#if DASHBOARD
-            var tstart = Stopwatch.GetTimestamp();
-            var tstop1 = tstart;
-            var lastWrittenValue = 0;
-            int count = 0;
-#endif
-
             Span<byte> value = stackalloc byte[kValueSize];
             ref SpanByte _value = ref SpanByte.Reinterpret(value);
 
@@ -514,20 +421,6 @@ namespace Tsavorite.benchmark
 
                         uContext.Upsert(ref SpanByte.Reinterpret(ref init_keys_[idx]), ref _value, Empty.Default);
                     }
-#if DASHBOARD
-                count += (int)kChunkSize;
-
-                //Check if stats collector is requesting for statistics
-                if (writeStats[thread_idx])
-                {
-                    var tstart1 = tstop1;
-                    tstop1 = Stopwatch.GetTimestamp();
-                    threadThroughput[thread_idx] = (count - lastWrittenValue) / ((tstop1 - tstart1) / freq);
-                    lastWrittenValue = count;
-                    writeStats[thread_idx] = false;
-                    statsWritten[thread_idx].Set();
-                }
-#endif
                 }
                 uContext.CompletePending(true);
             }
@@ -576,70 +469,6 @@ namespace Tsavorite.benchmark
 
             bContext.CompletePending(true);
         }
-
-#if DASHBOARD
-        int measurementInterval = 2000;
-        bool measureLatency;
-        bool[] writeStats;
-        private EventWaitHandle[] statsWritten;
-        double[] threadThroughput;
-        double[] threadAverageLatency;
-        double[] threadMaximumLatency;
-        long[] threadProgress;
-        double freq;
-
-        void DoContinuousMeasurements()
-        {
-            double totalThroughput, totalLatency, maximumLatency;
-            double totalProgress;
-            int ver = 0;
-
-            using (var client = new WebClient())
-            {
-                while (!done)
-                {
-                    ver++;
-
-                    Thread.Sleep(measurementInterval);
-
-                    totalProgress = 0;
-                    totalThroughput = 0;
-                    totalLatency = 0;
-                    maximumLatency = 0;
-
-                    for (int i = 0; i < threadCount; i++)
-                    {
-                        writeStats[i] = true;
-                    }
-
-
-                    for (int i = 0; i < threadCount; i++)
-                    {
-                        statsWritten[i].WaitOne();
-                        totalThroughput += threadThroughput[i];
-                        totalProgress += threadProgress[i];
-                        if (measureLatency)
-                        {
-                            totalLatency += threadAverageLatency[i];
-                            if (threadMaximumLatency[i] > maximumLatency)
-                            {
-                                maximumLatency = threadMaximumLatency[i];
-                            }
-                        }
-                    }
-
-                    if (measureLatency)
-                    {
-                        Console.WriteLine("{0} \t {1:0.000} \t {2} \t {3} \t {4} \t {5}", ver, totalThroughput / (double)1000000, totalLatency / threadCount, maximumLatency, store.Log.TailAddress, totalProgress);
-                    }
-                    else
-                    {
-                        Console.WriteLine("{0} \t {1:0.000} \t {2} \t {3}", ver, totalThroughput / (double)1000000, store.Log.TailAddress, totalProgress);
-                    }
-                }
-            }
-        }
-#endif
 
         #region Load Data
